Refactor core package: move agg logic to core/agg subpackage
