+++ core/agg/agg_parsing_test_utils.go
@@ -1,4 +1,5 @@
-package core
+// Package agg has aggregation logic for Git activity data.
+package agg
 
 import (
 	"fmt"
