+++ core/folders_test.go
@@ -3,6 +3,7 @@ package core
 import (
 	"testing"
 
+	"github.com/huangsam/hotspot/core/agg"
 	"github.com/huangsam/hotspot/internal/contract"
 	"github.com/huangsam/hotspot/schema"
 	"github.com/stretchr/testify/assert"
@@ -49,7 +50,7 @@ func TestAggregateAndScoreFolders(t *testing.T) {
 		Mode: schema.HotMode,
 	}
 
-	result := aggregateAndScoreFolders(cfg, fileResults)
+	result := agg.AggregateAndScoreFolders(cfg, fileResults)
 
 	// Should have 2 folders: src/, tests/ (root is skipped when PathFilter is empty)
 	assert.Len(t, result, 2)
@@ -118,7 +119,7 @@ func TestAggregateAndScoreFolders_WithPathFilter(t *testing.T) {
 		Mode:       schema.HotMode,
 	}
 
-	result := aggregateAndScoreFolders(cfg, fileResults)
+	result := agg.AggregateAndScoreFolders(cfg, fileResults)
 
 	// PathFilter doesn't affect folder aggregation - should still have both folders
 	assert.Len(t, result, 2)
@@ -128,7 +129,7 @@ func TestAggregateAndScoreFolders_EmptyInput(t *testing.T) {
 	fileResults := []schema.FileResult{}
 	cfg := &contract.Config{Mode: schema.HotMode}
 
-	result := aggregateAndScoreFolders(cfg, fileResults)
+	result := agg.AggregateAndScoreFolders(cfg, fileResults)
 
 	assert.Empty(t, result)
 }
@@ -149,7 +150,7 @@ func TestAggregateAndScoreFolders_SingleFileInRoot(t *testing.T) {
 		Mode: schema.HotMode,
 	}
 
-	result := aggregateAndScoreFolders(cfg, fileResults)
+	result := agg.AggregateAndScoreFolders(cfg, fileResults)
 
 	// Should not include root folder when PathFilter is empty
 	assert.Empty(t, result)
@@ -188,7 +189,7 @@ func TestAggregateAndScoreFolders_OwnerCalculation(t *testing.T) {
 		Mode: schema.HotMode,
 	}
 
-	result := aggregateAndScoreFolders(cfg, fileResults)
+	result := agg.AggregateAndScoreFolders(cfg, fileResults)
 
 	assert.Len(t, result, 1)
 	folder := result[0]
@@ -216,7 +217,7 @@ func TestAggregateAndScoreFolders_NoOwners(t *testing.T) {
 		Mode: schema.HotMode,
 	}
 
-	result := aggregateAndScoreFolders(cfg, fileResults)
+	result := agg.AggregateAndScoreFolders(cfg, fileResults)
 
 	assert.Len(t, result, 1)
 	folder := result[0]
